fix: Don't leak path fds to the child process
A command such as /bin/ls >> a.txt would leak the file descriptor
that refers to a.txt to the child process. The child has no way
of knowing about it.

This CL resolves this by closing any opened fd when we fork and are
in the child's code path.